requirements.txt
matplotlib
Pillow
opencv-python
numpy
requests